Restrict synchronization scope in IdentityHashCode
Since `InterceptorOfGlobalMethods$IdentityHashCode` relies on a
`WeakIdentityHashMap` instance for its bookeeping, `synchronized`
blocks involving accesses to said instance can cause a deadlock by
maintaining ownership of the `IdentityHashCode` instance monitor. To
get around this, we restrict usage of implicit locking only to fetching
a value from the (possibly shared) identity generator, and otherwise
rely on the underlying `AbstractQueuedSynchronizer` for safe accesses to
`WeakIdentityHashMap`.